+++ R/rayrender.R
@@ -6,10 +6,9 @@
 #' @examples
 #'   \donttest{
 #'     if (require("rayrender")) {
-#'         cfg <- pp_cfg()
+#'         cfg <- game_systems("sans3d")$piecepack
 #'         render_scene(piece("tile_face", suit = 3, rank = 3, cfg = cfg))
 #'         render_scene(piece("coin_back", suit = 4, rank = 2, cfg = cfg))
-#'         render_scene(piece("saucer_back", suit = 1, cfg = cfg))
 #'         render_scene(piece("pawn_face", suit = 2, cfg = cfg))
 #'     }
 #'   }