+++ R/rgl.R
@@ -12,11 +12,11 @@
 #' @examples
 #' if ((Sys.getenv("TRAVIS") == "") && require("rgl")) {
 #'     open3d()
-#'     cfg <- pp_cfg()
-#'     piece3d("tile_face", suit = 3, rank = 3, cfg = cfg, x = 0, y = 0, z = 0)
-#'     piece3d("coin_back", suit = 4, rank = 2, cfg = cfg, x = 2, y = 0, z = 0)
-#'     piece3d("saucer_back", suit = 1, cfg = cfg, x = 2, y = 2, z=-2)
-#'     piece3d("pawn_face", suit = 2, cfg = cfg, x = 1, y = 1, z = 2)
+#'     cfg <- game_systems("sans3d")$piecepack
+#'     piece3d("tile_back", suit = 3, rank = 3, cfg = cfg, x = 0, y = 0, z = 0)
+#'     piece3d("coin_back", suit = 4, rank = 2, cfg = cfg, x = 0.5, y = 0.5, z = 0.25)
+#'     piece3d("pawn_face", suit = 1, cfg = cfg, x = -0.5, y = 0.5, z = 0.25)
+#'     piece3d("pyramid_top", suit = 2, rank = 3, cfg = cfg, x = 0.5, y = -0.5, z = 0.5)
 #' }
 #' @export
 #' @seealso See \code{\link[rgl]{rgl-package}} for more information about the \code{rgl} package.