+++ NEWS.md
@@ -9,7 +9,7 @@ Bug fixes and minor improvements
 * `AA_to_R()` now accepts an optional `axis_z` argument and automatically normalizes the axis vector.
   In particular `R_to_AA(AA_to_R(...))` can now be used to normalize alternative axis-angle parameterizations
   with non-unit axis vectors and/or negative `axis_z` values.
-* Improved drawing of oblique projection edges for the "oval" shape (#212).
+* Improved drawing of oblique projection edges for the "oval" shape (#212) and the "halma" shape.
 * `is_color_invisible()` will now correctly classify as "invisible" colors with an alpha channel value set to 0.
 
 piecepackr 1.4.1